Use postgres instead of sqlite
Sqlite is a bit too restrictive in terms of changing the schema after
the initial migration. You can't add a column with a non-constant default
to an already created table, so you can't add a timestamp column with a
default set by `now`, you have to recreate the table from scratch and migrate
the data. You also cannot add an explicit primary key to an already
created table.