Add integration test of status endpoint
For integration testing the server I would like to spawn a server
instance in the test and then make a request to it.
But starting a task with a server started by warp::Server::run is possibly impossible to do
at the moment. I get the error
28 |         rt.spawn(setup_server(addr));
   |            ^^^^^ implementation of `warp::reply::Reply` is not general enough
   |
   = note: `&'0 str` must implement `warp::reply::Reply`, for any lifetime `'0`...
   = note: ...but `warp::reply::Reply` is actually implemented for the type `&'static str`

I could get it to work by using Server::bind.