Validate oauth2 provider issued tokens
By setting the `auth-type` query parameter to "oauth2" the provided
access token is validated as a self-encoded and signed access token.
This requires that the url for the oauth2 provider is given in the
OAUTH2_PROVIDER_BASE_URL environment variable so that the signing keys
can be fecthed there. This should be made optional since signing the
tokens isn't required (as far as I understand) so some providers might
not do it.