Use Arc for passing the db pool to the handlers
Wrap the pool object in a reference counter to avoid cloning the actual object every time it
is passed to a handler. The warp filters cannot take &-references but an Arc can be used to
avoid cloning the actual object. This way only the Arc is cloned on every call to .clone().
All the different Arcs will still point to the same allocation on the heap.